examples/multi_agent_human_chat/requirements.txt
eggai==0.1.44
python-dotenv==1.0.1
fastapi==0.115.6
dspy-ai==2.5.43
psutil==7.0.0
websockets==14.1
guardrails-ai==0.6.2
RAGatouille==0.0.9
uvicorn==0.34.0
openlit==1.33.10
opentelemetry-sdk==1.30.0
python-json-logger==2.0.7
colorlog==6.8.2
pydantic-settings==2.4.0
colbert-ai==0.2.21
transformers[torch]==4.49.0
sentence-transformers==4.0.2
numpy==2.2.4
scikit-learn==1.6.1
mlflow==2.21.3